Using mapdb with catbref patches, specifically a patch that uses thread IDs for BTreeMap locks instead of thread references which should reduce deadlocks.
For DBMaps that have a List or a Map as their value types,
introduce a shim subclass that deals with adding and removing sub-values
from the value-List/value-Map so that orphaning works correctly.
(This is mostly for name storage, blogs, etc.)
Also tidy descending index offset and get/set/contains/delete code.

Controller.java:
Re-rebuild name storage and blogpost maps in light of DBMap orphan updates.
Try to re-use data, and dataBak, directories instead of attempting to wipe them.
When shutting down, stop the block processor/synchronizer first as this can delay the block generator from shutting down.
Also add shutdown progress messages to be displayed via the improved ClosingDialog.

New ClosingDialog which calls Controller.stopAll() in a worker thread and shows progress messages.

Fix BlockMap so initial lastBlockSignature is correctly set to null with a new DB.
(Introduced in commit 9760b75)

Update BlogPostMap, HashtagPostMap, PostCommentMap, OrphanNameStorageHelperMap, SharedPostsMap to use new DBListValueMap subclass shim.
Update NameStorageMap, OrphanNameStorageMap to use new DBMapValueMap subclass shim.

Fix lots of unit tests: BlockTests, DatabaseTests

NOTE:
difflib-1.3.0 produces/applies empty-context unified diffs incorrectly.
See issue #20 at https://github.com/dnaumenko/java-diff-utils/issues/20

+++ Qora/src/utils/SimpleFileVisitorForRecursiveFolderDeletion.java
@@ -7,19 +7,23 @@ import java.nio.file.SimpleFileVisitor;
 import java.nio.file.attribute.BasicFileAttributes;
 
 public class SimpleFileVisitorForRecursiveFolderDeletion extends SimpleFileVisitor<Path> {
+	private Path excludeDir = null;
+
+	public SimpleFileVisitorForRecursiveFolderDeletion(Path excludeDir) {
+		this.excludeDir = excludeDir;
+	}
 
 	@Override
-	   public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-		   java.nio.file.Files.delete(file);
-		   return FileVisitResult.CONTINUE;
-	   }
+	public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
+		java.nio.file.Files.delete(file);
+		return FileVisitResult.CONTINUE;
+	}
 
-	
-	   @Override
-	   public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-		   java.nio.file.Files.delete(dir);
-		   return FileVisitResult.CONTINUE;
-	   }
+	@Override
+	public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+		if (excludeDir != null && dir != excludeDir)
+			java.nio.file.Files.delete(dir);
 
-	
+		return FileVisitResult.CONTINUE;
+	}
 }
